pyproject.toml
[project]
name = "customchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Daniel Molitor", email = "molitdj97@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "plotnine>=0.15.2",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "shiny>=1.5.1",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
